requirements.txt
pydantic-ai==1.41.0
a2a-sdk==0.3.22
allure-python-commons==2.15.2
python-dotenv==1.2.1
gunicorn==23.0.0
google-cloud-logging==3.13.0
fastapi==0.128.0
httpx==0.28.1
python-dateutil==2.9.0.post0
requests==2.32.5
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio==1.3.0
qdrant-client[grpc]==1.16.2
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
PyJWT==2.10.1
jira==3.10.5
